requirements.txt
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.1.0
entrypoints==0.3
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
nltk==3.6.5
numpy==1.21.2
pandas==1.3.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3
prompt-toolkit==3.0.20
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pywin32==301
pyzmq==22.3.0
regex==2021.10.8
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
setuptools==58.2.0
six==1.16.0
soupsieve==2.2.1
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
wheel==0.37.0
wordcloud==1.8.1
